pyproject.toml
# Refers to:
# - https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
# - https://packaging.python.org/en/latest/specifications/platform-compatibility-tags/
# - https://docs.python.org/3/c-api/stable.html
# - https://github.com/scikit-build/scikit-build-core
# - https://github.com/wjakob/nanobind/blob/master/docs/api_cmake.rst

[build-system]
requires = ["scikit-build-core", "nanobind"]
build-backend = "scikit_build_core.build"

[project]
name = "hybridsearch_embedded_sdk"
version = "0.6.0.dev3"
requires-python = ">=3.10"
dependencies = [
    "sqlglot~=11.7.0",
    "pydantic~=2.9.0",
    "thrift~=0.20.0",
    "setuptools~=75.2.0",
    "pytest~=8.3.0",
    "pandas~=2.2.0",
    "numpy~=1.26.0",
    "pyarrow~=17.0.0",
    "polars-lts-cpu~=1.9.0",
    "openpyxl~=3.1.0",
    "requests~=2.32.0"
]
description = "hybridsearch"
readme = "README.md"

[tool.scikit-build]
minimum-version = "0.4"
cmake.targets = ["embedded_hybridsearch_ext"]
cmake.build-type = "Release"
build-dir = "build/{build_type}"

wheel.py-api = "cp310"
wheel.packages = ["python/hybridsearch_embedded"]

[tool.pytest.ini_options]
addopts = "--strict-markers"
markers = [
    "L0",
    "L1",
    "L2",
    "L3",
    "complex",
    "slow",
    "nightly",
]

filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::ResourceWarning",
    "ignore:pkg_resources is deprecated as an API:DeprecationWarning",
    # "ignore::thrift.transport.TTransport",
    # note the use of single quote below to denote "raw" strings in TOML
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]